test(mitm-addon): split firewall connector tests
